Implemented BaseClip class and inheritance.
Used the idea of BaseClip by Christian to implement a common base class where all types of clips can inherit. All clips in BaseClip are contained within a CompositeVideoFile, so there is no conflict with text methods. All can be cropped, resized, and have text added to them, and added to a finalVideo object, which uses a similar inheritance to BaseClip.